# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * IslamNashra API - AI-generated Islamic news platform
 * OpenAPI spec version: 0.1.0
 */

export interface Post {
  id: string;
  title: string;
  body: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  hasImage: boolean;
  significanceScore: number;
  sourceNote: string;
  publishedAt: Date;
  expiresAt: Date;
  isBreaking: boolean;
  createdAt: Date;
}