# lib/api-zod/src/generated/types/preferenceInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * IslamNashra API - AI-generated Islamic news platform
 * OpenAPI spec version: 0.1.0
 */

export interface PreferenceInput {
  deviceId: string;
  /** @nullable */
  pushToken?: string | null;
  followedCategories?: string[];
  notificationsEnabled?: boolean;
}